Style sheet now applied to PdmUiTreeViewEditor.
Move some classes from PdmUiTreeViewEditor.cpp to the header file in order to make them available in RiuGuiTheme to adjust the style of tree views and keep the drag&drop styling. 
Removed getStyleSheet function.
